+++ perfume-api/src/main/java/io/perfume/api/review/application/in/comment/GetReviewCommentsUseCase.java
@@ -1,4 +1,4 @@
-package io.perfume.api.review.application.in;
+package io.perfume.api.review.application.in.comment;
 
 import dto.repository.CursorPagination;
 import io.perfume.api.review.application.in.dto.GetReviewCommentsCommand;
