+++ perfume-api/src/main/java/io/perfume/api/review/application/in/review/CreateReviewUseCase.java
@@ -1,4 +1,4 @@
-package io.perfume.api.review.application.in;
+package io.perfume.api.review.application.in.review;
 
 import io.perfume.api.review.application.in.dto.CreateReviewCommand;
 import io.perfume.api.review.application.in.dto.ReviewResult;
