+++ perfume-api/src/main/java/io/perfume/api/review/application/in/review/GetReviewCountUseCase.java
@@ -1,4 +1,4 @@
-package io.perfume.api.review.application.in;
+package io.perfume.api.review.application.in.review;
 
 public interface GetReviewCountUseCase {
