+++ perfume-api/src/main/java/io/perfume/api/review/application/in/tag/DeleteReviewTagUseCase.java
@@ -1,4 +1,4 @@
-package io.perfume.api.review.application.in;
+package io.perfume.api.review.application.in.tag;
 
 import java.time.LocalDateTime;
