+++ perfume-api/src/main/java/io/perfume/api/review/application/in/tag/GetReviewTagUseCase.java
@@ -1,4 +1,4 @@
-package io.perfume.api.review.application.in;
+package io.perfume.api.review.application.in.tag;
 
 import io.perfume.api.review.application.in.dto.ReviewTagResult;
 import java.util.List;
