+++ perfume-api/src/main/java/io/perfume/api/review/application/in/thumbnail/AddReviewThumbnailUseCase.java
@@ -1,4 +1,4 @@
-package io.perfume.api.review.application.in;
+package io.perfume.api.review.application.in.thumbnail;
 
 import io.perfume.api.review.domain.ReviewThumbnail;
 import java.time.LocalDateTime;
